refactor(states): reduce the number of files
