+++ src/lib.rs
@@ -8,4 +8,5 @@ pub extern crate websocket;
 
 extern crate rsc2_pb;
 
-pub mod states;
+pub mod engine;
+pub(crate) mod states;
